+++ frontend/src/boot/i18n.ts
@@ -5,7 +5,7 @@ import messages from "../i18n";
 import numberFormats from "../i18n/numberFormats";
 
 export const i18n = createI18n({
-  locale: "en-AU",
+  locale: navigator.language,
   fallbackLocale: "en-AU",
   numberFormats,
   messages,