+++ frontend/src/i18n/numberFormats.ts
@@ -1,12 +1,20 @@
 export default {
   "en-AU": {
     currency: {
-      style: "currency", currency: "AUD",
+      style: "currency",
+      currency: "AUD",
     },
   },
   "en-US": {
     currency: {
-      style: "currency", currency: "USD",
+      style: "currency",
+      currency: "USD",
+    },
+  },
+  "en-GB": {
+    currency: {
+      style: "currency",
+      currency: "GBP",
     },
   },
 };