+++ memberportal/membermatters/__init__.py
@@ -1,0 +1,3 @@
+from .celeryapp import app as celery_app
+
+__all__ = ("celery_app",)